Created the leadership change test.  It works!
To verify the server is behaving correctly, I had to move the type
information out of the server and into a shared type file that is used
in both the server and (symlinked to) the client.  This lets the
clients send requests to the server and then validate that the
servers' states are what they should be.

Run the tests with:

$ cd client; go test